Split utils into modules grouped by theme
utils.py had grown to 1442 lines and 23 unrelated functions, which made it hard
to read and to find anything in. The functions move into modules named after
what they do: datasets, distributions, glm, dispersions, stats, plotting, and
misc for the three that fit nowhere else.

The function bodies are unchanged and the dependencies between them were already
shallow enough that no module imports another cyclically. utils.py is gone
rather than left as a re-export shim, so `from pydeseq2.utils import ...` has to
become an import from the new module. Every caller in the package, the tests and
the examples is updated, as is the API documentation.

Closes #182
