+++ docs/source/api/docstrings/pydeseq2.datasets.rst
@@ -1,0 +1,15 @@
+﻿pydeseq2.datasets
+=================
+
+.. automodule:: pydeseq2.datasets
+
+
+
+
+
+
+   .. rubric:: Functions
+
+   .. autosummary::
+
+      load_example_data
