+++ docs/source/api/docstrings/pydeseq2.dispersions.rst
@@ -1,0 +1,21 @@
+﻿pydeseq2.dispersions
+====================
+
+.. automodule:: pydeseq2.dispersions
+
+
+
+
+
+
+   .. rubric:: Functions
+
+   .. autosummary::
+
+      dispersion_trend
+      fit_alpha_mle
+      fit_moments_dispersions
+      fit_rough_dispersions
+      robust_method_of_moments_disp
+      trimmed_cell_variance
+      trimmed_variance
