+++ docs/source/api/docstrings/pydeseq2.distributions.rst
@@ -1,0 +1,18 @@
+﻿pydeseq2.distributions
+======================
+
+.. automodule:: pydeseq2.distributions
+
+
+
+
+
+
+   .. rubric:: Functions
+
+   .. autosummary::
+
+      dnb_nll
+      nb_nll
+      nbinomFn
+      nbinomGLM
