+++ docs/source/api/docstrings/pydeseq2.glm.rst
@@ -1,0 +1,17 @@
+﻿pydeseq2.glm
+============
+
+.. automodule:: pydeseq2.glm
+
+
+
+
+
+
+   .. rubric:: Functions
+
+   .. autosummary::
+
+      fit_lin_mu
+      irls_solver
+      wald_test
