+++ docs/source/api/docstrings/pydeseq2.misc.rst
@@ -1,0 +1,17 @@
+﻿pydeseq2.misc
+=============
+
+.. automodule:: pydeseq2.misc
+
+
+
+
+
+
+   .. rubric:: Functions
+
+   .. autosummary::
+
+      get_num_processes
+      n_or_more_replicates
+      test_valid_counts
