+++ docs/source/api/docstrings/pydeseq2.plotting.rst
@@ -1,0 +1,16 @@
+﻿pydeseq2.plotting
+=================
+
+.. automodule:: pydeseq2.plotting
+
+
+
+
+
+
+   .. rubric:: Functions
+
+   .. autosummary::
+
+      make_MA_plot
+      make_scatter
