+++ docs/source/api/docstrings/pydeseq2.stats.rst
@@ -1,0 +1,17 @@
+﻿pydeseq2.stats
+==============
+
+.. automodule:: pydeseq2.stats
+
+
+
+
+
+
+   .. rubric:: Functions
+
+   .. autosummary::
+
+      lowess
+      mean_absolute_deviation
+      trimmed_mean
