--- docs/source/api/docstrings/pydeseq2.utils.rst
@@ -1,34 +1,0 @@
-﻿pydeseq2.utils
-==============
-
-.. automodule:: pydeseq2.utils
-
-
-
-
-
-
-
-   .. rubric:: Functions
-
-   .. autosummary::
-
-      dispersion_trend
-      dnb_nll
-      fit_alpha_mle
-      fit_lin_mu
-      fit_moments_dispersions
-      fit_rough_dispersions
-      get_num_processes
-      irls_solver
-      load_example_data
-      make_scatter
-      nb_nll
-      nbinomFn
-      nbinomGLM
-      robust_method_of_moments_disp
-      test_valid_counts
-      trimmed_cell_variance
-      trimmed_mean
-      trimmed_variance
-      wald_test
