+++ src/pydeseq2/distributions.py
@@ -1,0 +1,338 @@
+"""Negative binomial log-likelihoods and their optimization."""
+
+import numpy as np
+from scipy.optimize import minimize  # type: ignore
+from scipy.special import gammaln  # type: ignore
+from scipy.special import polygamma  # type: ignore
+
+from pydeseq2.grid_search import grid_fit_shrink_beta
+
+
+def nb_nll(
+    counts: np.ndarray, mu: np.ndarray, alpha: float | np.ndarray
+) -> float | np.ndarray:
+    r"""Neg log-likelihood of a negative binomial of parameters ``mu`` and ``alpha``.
+
+    Mathematically, if ``counts`` is a vector of counting entries :math:`y_i`
+    then the likelihood of each entry :math:`y_i` to be drawn from a negative
+    binomial :math:`NB(\mu, \alpha)` is [1]
+
+    .. math::
+        p(y_i | \mu, \alpha) = \frac{\Gamma(y_i + \alpha^{-1})}{
+            \Gamma(y_i + 1)\Gamma(\alpha^{-1})
+        }
+        \left(\frac{1}{1 + \alpha \mu} \right)^{1/\alpha}
+        \left(\frac{\mu}{\alpha^{-1} + \mu} \right)^{y_i}
+
+    As a consequence, assuming there are :math:`n` entries,
+    the total negative log-likelihood for ``counts`` is
+
+    .. math::
+        \ell(\mu, \alpha) = \frac{n}{\alpha} \log(\alpha) +
+            \sum_i \left \lbrace
+            - \log \left( \frac{\Gamma(y_i + \alpha^{-1})}{
+            \Gamma(y_i + 1)\Gamma(\alpha^{-1})
+        } \right)
+        + (\alpha^{-1} + y_i) \log (\alpha^{-1} + \mu)
+        - y_i \log \mu
+            \right \rbrace
+
+    This is implemented in this function.
+
+    Parameters
+    ----------
+    counts : ndarray
+        Observations.
+
+    mu : ndarray
+        Mean of the distribution :math:`\mu`.
+
+    alpha : float or ndarray
+        Dispersion of the distribution :math:`\alpha`,
+        s.t. the variance is :math:`\mu + \alpha \mu^2`.
+
+    Returns
+    -------
+    float or ndarray
+        Negative log likelihood of the observations counts
+        following :math:`NB(\mu, \alpha)`.
+
+    Notes
+    -----
+    [1] https://en.wikipedia.org/wiki/Negative_binomial_distribution
+    """
+    n = len(counts)
+    alpha_neg1 = 1 / alpha
+    logbinom = gammaln(counts + alpha_neg1) - gammaln(counts + 1) - gammaln(alpha_neg1)
+    if hasattr(alpha, "__len__") and len(alpha) > 1:
+        return (
+            alpha_neg1 * np.log(alpha)
+            - logbinom
+            + (counts + alpha_neg1) * np.log(mu + alpha_neg1)
+            - (counts * np.log(mu))
+        ).sum(axis=0)
+    else:
+        return (
+            n * alpha_neg1 * np.log(alpha)
+            + (
+                -logbinom
+                + (counts + alpha_neg1) * np.log(alpha_neg1 + mu)
+                - counts * np.log(mu)
+            ).sum()
+        )
+
+
+def dnb_nll(counts: np.ndarray, mu: np.ndarray, alpha: float) -> float:
+    r"""Gradient of the negative log-likelihood of a negative binomial.
+
+    Unvectorized.
+
+    Parameters
+    ----------
+    counts : ndarray
+        Observations.
+
+    mu : float
+        Mean of the distribution.
+
+    alpha : float
+        Dispersion of the distribution,
+        s.t. the variance is :math:`\mu + \alpha\mu^2`.
+
+    Returns
+    -------
+    float
+        Derivative of negative log likelihood of NB w.r.t. :math:`\alpha`.
+    """
+    alpha_neg1 = 1 / alpha
+    ll_part = (
+        alpha_neg1**2
+        * (
+            polygamma(0, alpha_neg1)
+            - polygamma(0, counts + alpha_neg1)
+            + np.log(1 + mu * alpha)
+            + (counts - mu) / (mu + alpha_neg1)
+        ).sum()
+    )
+
+    return -ll_part
+
+
+def nbinomFn(
+    beta: np.ndarray,
+    design_matrix: np.ndarray,
+    counts: np.ndarray,
+    size: np.ndarray,
+    offset: np.ndarray,
+    prior_no_shrink_scale: float,
+    prior_scale: float,
+    shrink_index: int = 1,
+) -> float:
+    """Return the NB negative likelihood with apeGLM prior.
+
+    Use for LFC shrinkage.
+
+    Parameters
+    ----------
+    beta : ndarray
+        2-element array: intercept and LFC coefficients.
+
+    design_matrix : ndarray
+        Design matrix.
+
+    counts : ndarray
+        Raw counts.
+
+    size : ndarray
+        Size parameter of NB family (inverse of dispersion).
+
+    offset : ndarray
+        Natural logarithm of size factor.
+
+    prior_no_shrink_scale : float
+        Prior variance for the intercept.
+
+    prior_scale : float
+        Prior variance for the intercept.
+
+    shrink_index : int
+        Index of the LFC coordinate to shrink. (default: ``1``).
+
+    Returns
+    -------
+    float
+        Sum of the NB negative likelihood and apeGLM prior.
+    """
+    num_vars = design_matrix.shape[-1]
+
+    shrink_mask = np.zeros(num_vars)
+    shrink_mask[shrink_index] = 1
+    no_shrink_mask = np.ones(num_vars) - shrink_mask
+
+    xbeta = design_matrix @ beta
+    prior = (
+        (beta * no_shrink_mask) ** 2 / (2 * prior_no_shrink_scale**2)
+    ).sum() + np.log1p((beta[shrink_index] / prior_scale) ** 2)
+
+    nll = (
+        counts * xbeta - (counts + size) * np.logaddexp(xbeta + offset, np.log(size))
+    ).sum(axis=0)
+
+    return prior - nll
+
+
+def nbinomGLM(
+    design_matrix: np.ndarray,
+    counts: np.ndarray,
+    size: np.ndarray,
+    offset: np.ndarray,
+    prior_no_shrink_scale: float,
+    prior_scale: float,
+    optimizer="L-BFGS-B",
+    shrink_index: int = 1,
+) -> tuple[np.ndarray, np.ndarray, bool]:
+    """Fit a negative binomial MAP LFC using an apeGLM prior.
+
+    Only the LFC is shrinked, and not the intercept.
+
+    Parameters
+    ----------
+    design_matrix : ndarray
+        Design matrix.
+
+    counts : ndarray
+        Raw counts.
+
+    size : ndarray
+        Size parameter of NB family (inverse of dispersion).
+
+    offset : ndarray
+        Natural logarithm of size factor.
+
+    prior_no_shrink_scale : float
+        Prior variance for the intercept.
+
+    prior_scale : float
+        Prior variance for the LFC parameter.
+
+    optimizer : str
+        Optimizing method to use in case IRLS starts diverging.
+        Accepted values: 'L-BFGS-B', 'BFGS' or 'Newton-CG'. (default: ``'Newton-CG'``).
+
+    shrink_index : int
+        Index of the LFC coordinate to shrink. (default: ``1``).
+
+    Returns
+    -------
+    beta: ndarray
+        2-element array, containing the intercept (first) and the LFC (second).
+
+    inv_hessian: ndarray
+        Inverse of the Hessian of the objective at the estimated MAP LFC.
+
+    converged: bool
+        Whether L-BFGS-B converged.
+    """
+    num_vars = design_matrix.shape[-1]
+
+    shrink_mask = np.zeros(num_vars)
+    shrink_mask[shrink_index] = 1
+    no_shrink_mask = np.ones(num_vars) - shrink_mask
+
+    beta_init = np.ones(num_vars) * 0.1 * (-1) ** (np.arange(num_vars))
+
+    # Set optimization scale
+    scale_cnst = nbinomFn(
+        np.zeros(num_vars),
+        design_matrix,
+        counts,
+        size,
+        offset,
+        prior_no_shrink_scale,
+        prior_scale,
+        shrink_index,
+    )
+    scale_cnst = np.maximum(scale_cnst, 1)
+
+    def f(beta: np.ndarray, cnst: float = scale_cnst) -> float:
+        # Function to optimize
+        return (
+            nbinomFn(
+                beta,
+                design_matrix,
+                counts,
+                size,
+                offset,
+                prior_no_shrink_scale,
+                prior_scale,
+                shrink_index,
+            )
+            / cnst
+        )
+
+    def df(beta: np.ndarray, cnst: float = scale_cnst) -> np.ndarray:
+        # Gradient of the function to optimize
+        xbeta = design_matrix @ beta
+        d_neg_prior = (
+            beta * no_shrink_mask / prior_no_shrink_scale** 2
+            + 2 * beta * shrink_mask / (prior_scale**2 + beta[shrink_index] ** 2)
+        )
+
+        d_nll = (
+            counts - (counts + size) / (1 + size * np.exp(-xbeta - offset))
+        ) @ design_matrix
+
+        return (d_neg_prior - d_nll) / cnst
+
+    def ddf(beta: np.ndarray, cnst: float = scale_cnst) -> np.ndarray:
+        # Hessian of the function to optimize
+        # Note: will only work if there is a single shrink index
+        xbeta = design_matrix @ beta
+        exp_xbeta_off = np.exp(xbeta + offset)
+        frac = (counts + size) * size * exp_xbeta_off / (size + exp_xbeta_off) ** 2
+        # Build diagonal
+        h11 = 1 / prior_no_shrink_scale**2
+        h22 = (
+            2
+            * (prior_scale**2 - beta[shrink_index] ** 2)
+            / (prior_scale**2 + beta[shrink_index] ** 2) ** 2
+        )
+
+        h = np.diag(no_shrink_mask * h11 + shrink_mask * h22)
+
+        return 1 / cnst * ((design_matrix.T * frac) @ design_matrix + np.diag(h))
+
+    res = minimize(
+        f,
+        beta_init,
+        jac=df,
+        hess=ddf if optimizer == "Newton-CG" else None,
+        method=optimizer,
+        options={
+            "ftol": 1e-8,
+            "gtol": 1e-8,
+        },
+    )
+
+    beta = res.x
+    converged = res.success
+
+    if not converged and num_vars == 2:
+        # If the solver failed, fit using grid search (slow)
+        # Only for single-factor analysis
+        beta = grid_fit_shrink_beta(
+            counts,
+            offset,
+            design_matrix,
+            size,
+            prior_no_shrink_scale,
+            prior_scale,
+            scale_cnst,
+            grid_length=60,
+            min_beta=-30,
+            max_beta=30,
+        )
+
+    inv_hessian = np.linalg.inv(ddf(beta, 1))
+
+    return beta, inv_hessian, converged
